Task 1: Ordering
This commit provides the following. A method of querying and relating
foreign key columns to their related counterparts in the following
format.
| Table Name | Column Name | Related Table Name | Related Column Name |

This code is currently in the `db` package. There is also some code there
to display the table information that is currently unused.

This commit also has the code for creating the order of the tables that
need to be filled in the `graph` pacakge. This allows the above foreign
key relationships to be used to identify the order in which the tables
related to a specific table need to be filled in order to maintain
proper fk (Foreign Key) relationships.

Some JSON files were also added to the repo to serve as visual examples.