Task 2: Testing
This commit includes all of the changes neccesary to enable unit testing
for the `ordering.go` file. These changes include creating the migration
files for each test case, the creation of the `ordering_test.go` file as
well as tampering with the `db.go` file.

The `ordering_test.go` file's main function is to test the `FindOrder`
function in the `ordering.go` file. In order to do this it uses the
golang migrate library to drop and rebuild table schemas using Go migrate.

As a result of the testing some changes were made to the `FindOrder`
function and the `ordering.go` file in general. An abridged version of
these changes are that the `Ordering` struct now has a reference to a
map of a table names. The `FindOrder` function now checks to see if the
given table name exists in the database before continuing. If the table
doesn't a new error type (MissingTableError) is returned.

`db.go` received a new function called `GetTableMap` which just returns
a map of all existing tables in a database. The file also no longer
imports from the `graph` package in order to not trigger a circular
dependency. As a result the `saveRelationships()` function was deleted.