Task 3: Making Parameters and Parsers
This commit contains all of the changes necessary for the creation of
the code enums needed to generate data as well as the
parsers needed to actually generate the INSERT queries.

In order to do this, a new package, `parameters`, needed to be created
which would contain the above logic. In order to support this new
package, changes needed to also be made to the `graph` and `db` packages.

Simple Changes:
- Map in the `db` package which maps the type of the column to the
  parser that handles that type. For example, the type INT32 would be
  mapped to the INT value since the integer parser handles all integer
  types.

- `db` package now exports the new `GetColumnMap()` function which returns a map
  of the column's name to it's type for a given table name.

- `db` package now exports the `GetTablePKMap()` function which returns all of
  the primary keys for all tables in the database (no use for this yet)

- New Migration files added for personal testing

- The `Ordering` struct now has an `Init` function allowing for its
  convenient use in other packages.

The parameters package mostly consists of 2 things `QueryWriter` and
`ColumnParser`.

`QueryWriter`: This struct is responsible for writting the INSERT
Queries needed to actually generate data in the database. It does that
by going through a queue of tables and then generating values for each
column through the use of a `ColumnParser`.

`ColumnParser`: `ColumnParser` is an interface with one method,
`ParseColumn()` which returns a string and an error. The `ParseColumn()`
function evaluates the column and uses codes found in `codes.go` to
generate the data.

The `parser.go` function also has a map that maps the column's
datatype to the Parser struct. For example, an INT column would be
mapped to the IntColumnParser`.

Two other structs were created to help simplify some changes. These
structs were the `table` struct which has the table's name as well as a
list of `column` structs. The `column` struct is a struct containing all
the information a parser would need to generate data for the column. It
also has a reference to it's parser. This is done so that the SEQ code
actually functions properly since it is tied to that column and that
column alone.

Other Changes:
- Made some JSON file to be used as a template for data generation