Task 4: Improve Algorithms
This task is about improving the efficiency of the algorithms used to
get the queries. As a result this commit affects both the `graph` and
`parameters` packages. The below changes were all changes to the
`Ordering` struct.

Instead of using the algorithm I came up with I realized it was better
to just use a topological sorting gotten from DFS(Depth First Search).
By doing this I can get the order of the nodes that I need to go through
in an efficient manner so long as database schema doesn't have a cycle.
The function responsible for that is called `Topological()` and it takes
in the table where we start the DFS at. It returns the order of the
tables in the form of a queue.

Now the problem became how to detect and report on cycles since my
previous solution to that problem was too simplistic. The answer to
that, just like the answer to most graph problems is to use DFS to look
for cycles. The functions that control this are the following,
`findCycles()`, `hasCyclesForTable()`, and `HasCycles()`.

`findCycles()`: is the actual DFS algorithm that runs DFS based off a
given starting table. This function doesn't check if the table given
exists since in it's actual use case it's always checked before this is
ran. This function returns a list of cycles that were found as well as a
map of all the tables visited.

`hasCyclesForTable()`: This function is just a mask for `findCycles()`
which just runs `findCycles()` but only returns the list of cycles.

`HasCycles()`: This function checks the entire database schema to see if
there are any cycles. It does this by running `findCycles()` for a given
table name that wasn't covered yet. It then returns the list of found
cycles.

Other Changes:
- `TopologicalNode` struct created, it's basically the nodes in a graph
  and it has all of the information needed for my implementation of DFS
- Function `GetTopologicalNodes()` function created which basically
  forms a `TopologicalNode` struct and maps it to the table's name. It
  then returns that map.
- Changed `Cyclic` error to now take in the list of cycles detected and
  output a string that shows all of the cycles.
- Slightly changed the error message for the `MissingTableError` struct
- Modified `QueryWriter`'s functions to be compatible with the new
  changes made in the `Ordering` struct.
- Added new schemas with cycles for testing. Manual for now but I'll
  automate it when I feel like it.
- Fixed the unit tests in the `ordering_test.go` files to work with the
  new code.