Task 5.1: Housekeeping and Bugfixing
This series of commits doesn't provide any new functionality or features, instead it focuses on renaming and refining existing code.

The following functions received name changes:
- `HasCycles()` --> `GetCycles()`
- `hasCyclesForTable()` --> `getCyclesForTable()`
- `FindOrder()` --> `GetOrder()`
- `CycleBreaking()` --> `GetCycleBreakingOrder()`

As a consequence all references to these functions needed to be edited.

a series of functions and structs became private because they really don't need to be used  outside their package.