Task 5.2: Further Refining the Code
This series of changes is meant to make the code more easily used as a
library. This involves helper functions to make a proper `Ordering` and
`QueryWriter` structs. The `CreateSuggestions()` function was moved back
to the `Ordering` struct because a `QueryWriter` should not be able to
be made if there is a cycle in the path for the given table.

Helper Functions Made:
- `NewOrdering()`: returns the pointer to a `Ordering` struct that has
  ran the `Init` function
- `NewQueryWriterFor()`: This function takes in a table name and returns
  the pointer to the `QueryWriter` and any errors
- `ChangeTableToWriteFor()`: Takes in a table name and reinitializes the
  `QueryWriter` struct. It returns any errors from the `Init()` call.

Other changes:
- `allRelations` in the `QueryWriter` struct is now private
- updated the main file