Task 5.3: Final Refinement
This series of changes are made to further prepare the code to be used as a     library. The first of these changes is to make every function in the `db`       package now rely a database pointer (`*sql.DB`) to work. This is so that users  can have those functions work on their own databases without having to fork overthe repo and edit my configurations.

As a consequence of this, both `Ordering` and `QueryWriter` structs now need a
database pointer to work properly and this is reflected in their `Init`
functions. These database pointers are private so as to stop them from being    used improperly.

The `GetCycleBreakingOrder()` function is now private since we wouldn't want a
user to give it faulty inputs and have it break.

The functions `GetSuggestionQueries()`  and  `GetAndResolveCycles()` function were added to mask the fact that `GetCycleBreakingOrder()` is now private.

- `GetSuggestionQueries()`: this function takes in nothing and returns a list of queries that, if ran, should fix any cycles detected.

- `GetAndResolveCycles()`: This function takes in nothing and creates a list of suggestion queries and runs them on the given database. It panics if a query
fails.

Finally, I added a new test case for a table that references itself.
Currently, this fails the test so this behavior will need to be fixed.