Task 7: CLI Development
This PR focuses on adding a CLI for this project. This CLI will serve a
simpler function and will allow users to not be restricted to Golang in
order to use this code. This CLI uses the cobra golang library for
support in this endeavor. This PR can be considered the initial commit
towards this vision.

Documentation regarding the commands will be, eventually, both in the
documentation of the CLI AND in a README with examples.

As a result of making the CLI some changes needed to be made for the
sake of the project. Firstly, the `Other` field in the `column` struct
is now a string value. This is because I want the library to support
templating and unlike in python Golang actually
has types I need to respect and I don't feel like dealing with any type
nonsense right now.

As a result of this change, the parsers all needed to have some tweaks
to accommodate that. This isn't too hard to change since there is no code
that supported using multiple values at once. This does mean however
that multi-value codes can not be implemented for the time being.

Another consequence of using JSON templates is validating them. As a
result, the following functions were created. `verifyTemplate()` is a
function that verifies if the JSON map is valid by doing all sorts of
comparisons. If the map is a valid template for the table, the function
returns true, otherwise it returns false.

The `updateTableMap()` updates the information in the table structs
according to the given map. This function is only ever called after the
validation has happened.

Finally, the `cmd` directory has all the code relating to the cobra
library.

Before I can continue working on the CLI however, I must first fix a bug
with the project.