Task 7.1: Fixing Compound Key Bug
There was a bug in which the foreign key relationship for compound keys
wasn't properly destroyed. For example, let's say table "E" had a
compound key reference (bref and bref2) only bref2 would be dropped
instead of both of them.

This was because due to an oversight with the inverse relationship map
that I created. This error is now fixed by creating a new string builder
and making the `appendDropBuilder()` function which creates the
formatted string.

Unlike the other string builder functions, the output of this function
is split to make the queries more readable. As a result, the last index
of the string array will always be empty string and as such, useless. A
solution to this would be to simply have a 1 line alter statement that
drops multiple columns at once. While possible, eventually, I will have
to extend the library to support multiple SQL dialects and they might
all support difference syntax for that and I don't feel like crying just
yet.