Task 8: CLI Feature Completion
This PR is focused on finalizing the features for the CLI portion of the
project. This mostly meant finishing up the templating functionality of
the project as well as just blanket optimizations. Here are the changes
summarized.

The `db` package received the `RunQueriesVerbose()` function which just
prints out a formatted query string before executing it.

The `ColumnParser` interface received a change to the `ParseColumn()`
function. Now, the function takes in the column it's meant to parse.
This was because I was an idiot and I forgot to implement a method to
connect the Column to the parser. After some trial and error, this was
essentially the only non-headache inducing way of fixing the problem.

A fun consequence of this is that I had to apply this to all of the
parsers... yay. Another consequence of this is that the column array in
the `table` struct is not an array of type `*column`. This is just so
that updating the column information from a template is actually
possible. As a result, I needed to tweak the `createTable()` function.