Task 11: Making Release Workflow
This commit is focused around creating the workflow for the binary
releases using both goreleaser and GitHub workflows. The binary is
created and signed in the `.goreleaser.yaml` file using `cosign`.
`cosign` signs the binary with the `cosign.key` file and this release is
published on GitHub. The workflow is triggered whenever a new version
tag is pushed to the repo which should be infrequent. The release also
builds the binary for different platforms like Windows and Linux.

The signing is done to deter the ever present threat of Windows Defender
flagging the binary as a trojan virus for whatever reason.

Further work can be done to add this CLI to package managers like `yum`
or `chocolatey` for even easier access for the end user. Currently,
since we have no users, that is not necessary.