Task 12: Refining Functions
This PR is for refining functions in the library and making them more
efficient/easier to use for the user. This PR also deletes some old
files served as a visual reference.

The main function that were streamlined was the `GetInverseRelationships()`
because it had data that was never going to be used for it's function.

Another aspect of this PR was to make every public function that returns
a linked list, return a string array instead. This is a quality of life
change for the end user since looping through an array is easier than
going through a linked list. This is mostly accomplished with the new
`Utils` package's function, `ListToStringArray()`. As a result of this
change, some logic for the functions needed to be changed.

As a result of this, some functions like `RunQueries()` and
`RunQueriesVerbose()` now take in a string array as an argument rather
than a linked list.

Finally, the last change was to change the `db.Query()` function call in
the `RunQueries()` and `RunQueriesVerbose()` to `db.Exec()` calls instead.
This is done because `Query()` is meant to be used with SELECT
statements. As a result, running a large amount of INSERT and DELETE
queries with this function results in the `too many users` error in postgres.
Another benefit to this is that it speeds up the code's execution by a
lot.

Others:
- `insertQueryQueue` and `deleteQueryQueue`is now private
- `GenerateEntries()` and `GenerateEntry()` now returns the string array
  representation of `insertQueryQueue` and `deleteQueryQueue`
- Database connections are closed after every CLI command using
  `db.Close()` and defer
- removing some unused code