Task 13: Creating Update palette + Upgrades
This commit is focused on creating a new palette, `update`, with the
`template` command. This palette is responsible for updating any data
that is generated by the CLI. For now, it is only used for updating
templates using the `template` command.

Due to this, a series of wider general changes were made in order to not
repeat work.

The `db` package received 2 new functions, `InitDB()` and `CloseDB()`.
- `InitDB()`: Takes in a connection string and returns the database
  connection alongside any errors that occurred.
- `CloseDB()`: Takes in a database connection and closes it. It also
  logs any errors that occurred while closing.

These functions were made because each palette had their respective
`InitDB()` function and a function to close a database connection.
This was just repeated work. To fix that, these functions in the `db`
package were created and will be imported by all palettes.

The `utils` package also received new functions. Because the logic for
updating a template was similar to making one, the package received the
`MakeTemplates()` and `makeTemplate()` functions.

- `MakeTemplates()`: Takes in a database connection and an array of
  tables and formats it into a map which is returned.
- `makeTemplate()`: Takes in a database connection and a table name.
  This is responsible for making the individual table's template map.

Finally, the last part of this commit is creating the `template` command
for the `update` palette. This command verifies the integrity of the
given template and overwrites it with an updated template. It also maps
entries from the old template to the new template.