Task 14: New Command `queries` Created
This commit implements the `queries` command under the `generate`
palette. The `queries` command generates and prints the insert queries
to an output file instead of running them like `entry` does.

While some more work needs to be done on this, refactoring the CLI
is the priority which is why this commit is going to be pretty bare
bones.