Task 15: Refactor CLI + bug fixes
This commit is about refactoring the CLI. Before, there were 3 command
palettes, `validate`, `generate` and `update`. After some thinking, it
was decided to move the former `template` command over to the newly
created `template` palette (formerly known as `update`).

Because of this, the following happened:

- `generate/template.go` --> `template/create.go`
- `update/update.go` --> `template/template.go`
- `update/template.go` --> `template/update.go`

This was followed by a series of optimizations using cobra's persistent
flags. Some minor changes were made to the individual commands but it
isn't notable enough to mention. These changes were mostly there to
ensure proper formatting for folder, and file names.

I also patched the dumbest bug I've ever created in `getFrequency()`
where I accidentally skipped the first index in the array which resulted
in table "b" never being part of the count. This ensured a random output
for the generate function as less effective tables were used to break
the cycle.

Finally, `utils` received the `TrimAndLowerString()` function because I
got tired of re-writing it. I also added 2 files to gitignore for my
own convenience.