Task 16: Improve Parsers + Others
This commit was for adding more types that are supported to the
`typeMap` in the `db.go` file as well as testing and improving upon
current `ColumnParsers`. As a result of this, various changes were
made to the `ColumnParsers`. A new column parser was also made,
`FloatColumnParser` which is responsible for creating values for
any value that is a floating point number.

Other changes:
- default regex for `VarcharColumnParser` got limited to 10 characters
  because of the large strings it kept dumping into the queries.

- Added a condition in the `processTable()` function which adds single
  quotes(' ') to non-numerical values, that way they are recognized by
  the database

- changed the SEQ code's function so that it starts at value 1 like what
  SERIAL does in Postgres.

- Capped the date string generated by the NOW code in `DateColumnParser`
  to 19 characters because the full string wasn't accepted as a valid
  timestamp.

- Created a realistic schema to test various types

- Created an error when an unsupported column type is parsed. This
  prevents a nil pointer error which could be confusing for the end
  user.

- the `dir` flag is no longer required in the `template create` command.