Task 17: Updating Parsers and Suggestions
This commit handles the problem of keeping the precision and length of
certain types consistent as well as updating parser outputs.

The first problem:
The first issue this commit tackles is a loss of data when dealing with
cyclical relationships. Say, for the sake of the argument, a column called
Description VARCHAR(100) is affected when dealing with a
cycle. The code will create a translation table and a new column called
(insert table name)_descriptions_ref VARCHAR will be made.
If you've noticed, the constraint of 100 characters is lost in the
translation table.

Of course, this requires a VARCHAR to be a primary key which is a
horrible idea. However, you never know what schemas this will run on so better
to deal with it now. This same problem applied to floating point types
like DECIMAL and NUMERIC as you can customize their precision and scale.

The second problem:
The second issue this commit tackles is handling constraints such as
length, precision and scale based on the column. This issue really
only applies to `VarcharColumnParser` and `FloatColumnParser`.
Luckily, the `schema` library that supports this project has the
ability to get that via the `ColumnTypes()` function that returns
the `sql.ColumnType` structs that contain all of the information required.
This information is then mapped to the column's name before being
returned by `GetRawColumnMap()`.

The `column` struct in `table.go` now takes this information as the
field, `ColumnDetails`. This new column information can now be referenced
by the parser when generating values. In the case of `FloatColumnParser`
it also explicitly casts the data to certain formats in the returned string
using the `::` operator in Postgres.

The reason for this type cast is that it is done automatically for
INSERT queries but not for DELETE queries. Meaning, the delete queries
will fail since the value they have no longer matches the value in the
database. For this reason, all values generated by the
`FloatColumnParser` is casted to it's appropriate type.

Other changes:
- Correct the default regex expression based on character constraints
- Correct the default range of a float based off of the precision and
  scale if there is one.