Task 19: Self-Reference Bug Fix
This commit focuses on addressing a bug with a table that references
itself as a foreign key. I haven't updated this code in centuries so it
isn't surprising that it lead to a bug. I should probably add a test
case for it when I feel like it. This bug was fixed. I also created a
new helper function to shorten the work done for `appendBuilder()`. This
new function is called `getColumnQuery()`. It does all of the processing
that `appendBuilder()` did and returns the query string.

The test case for the self referencing schema was also updated to have
multiple columns.