Task 20: Fixing Bug in Suggestions
Before this, the suggestions would erase ALL foreign key relationships
from the `problem` table. This is problematic behavior because if a
table is not part of a cycle, it should not but altered.

The way the program determines if a table is relevant is if it is a part
of the current cycle. If it is, it continues. If not, it ends.

Another change that was made was to also remove the foreign key column in the
referenced table if it references the problem table. This way, the
relationship is preserved only through the translation table.