Task 21: Fixing a whole lot of bugs
This task is focused on fixing a variety of edge cases and bugs in the
programs logic. As such, each section will describe a problem and how it
was solved in this commit.

The Way to Determine if a Key is Relevant to Fixing a Cycle was Wrong:
The previous logic assumed that to remove the relationship between 2
tables, the foreign key reference to the problem table had to be to a
`Primary Key`. This is false, the referenced key simply needs to be `Unique`
which, by definition includes `Primary Keys`. For example, an email is a
unique value, even if it isn't used as a primary key, it can be
referenced for a foreign key. To fix this problem, a function called
`getRelevantKeys()` was created which returns a list of the ACTUAL
referenced keys from the problem table. In the case of a self-reference
cycle, it returns a list of the foreign keys that reference it's own
table i.e. self-reference. The referenced table keys are always
it's primary keys so that section of the code received no changes.

Data Loss When Resolving Cycles:
Currently, the program creates the translation table and then drops the
foreign keys that reference the problem table. This causes the data in the
columns to be lost which, in most cases, is unacceptable behavior. To fix
this, before the column(s) is/are dropped, it is inserted into the new
translation table via a JOIN query using `INSERT INTO SELECT`. The
logic for this behavior is written in the `appendColumnBuilder()`
function. Once the data is moved, the column(s) can be safely dropped,
preserving the existing relationships between the tables. Each column is
also associated with their table in this query to prevent ambiguous column
errors using the following format, Table_Name.Column_Name.

Bloated and Unnecessary Complexity:
Beforehand, the logic for handling the queries of self-referencing
tables and the general case divided. This kept the query writting
functions (`appendBuilder()`, `AppendForeignKey()`, etc.) simple. The
problem however, was that it made the code difficult to read. Now, the
functions required to write the queries handle that logic, making the
code easier to read and streamlining the logic. These functions also
received general improvements and additions to handle the increased
complexity.

Logic for Determining Relevant Tables is Wrong:
The current logic for determining if the relationship between a table
needs to be removed (i.e. is relevant to getting rid of a cycle) is to
check if the table is in A cycle. This has been fixed by giving the
`getSuggestions()` function a map of the problem tables and a map of all
the relevant tables. We construct that map and return it in the
`getCycleBreakingOrder()` function with the help of the
`constructProblemTableMap()` function.

Lack of a proper Omni Test Case Database Schema:
Most of the work for this commit is focused around handling edge cases.
It would be good to have a schema that involves all those edge cases to
ensure that the program works as intended. This schema would also be
useful for future automated testing. As such, case 9 was created.
Further detail regarding what exactly this schema checks for will be
given in a later PR.

Considering how the program works, I believe none of these changes
should break the binary... Hopefully.

Other Changes:
- Removed the "ref" portion of the name for the new key in the table,
  the format tableName_ColumnName will be used and it ensures a unique key.
- Overall efficiency improvements in the code
- Standardized naming conventions (problemTable, refTable etc.)
- Got rid of unecessary input in functions

Glossary:
- `Problem Table`: Table that appears in the most cycles, getting rid of
  the relevant tables that reference the problem table solves the cycles.
  There can be multiple problem tables in any given database. By "solving" all
  problem tables, you get rid of all cycles.
- `Reference Table`: A table that references a problem table AND is in
  the same cycle as that same problem table. It is also sometimes referred
  to as a "relevant table".
- `Relevant Table`: A table is considered "relevant" if it is in the same
  cycle as a problem table.
- `Translation Table`: A table that is created which has the primary keys
  of the referenced table and the keys referenced from the problem
  table. This allows the relationship between the tables to be
  maintained while getting rid of the cycle.