Task 22: Add New 'Table' command to CLI
This commit is all the changes relating to implementing the new 'table'
command in the CLI. This command will be part of the 'validate' palette
and is used to see if the given table is involved in any cycles.

To support this, new functions were added to the `graph` package.

Function added to `helpers.go`:
- `getRelevantCycles()`: This function takes in a string and a
  linked list of cycle strings. It then removes any irrelevant cycles
  (cycles that don't contain the given string). It then returns
  the remaining cycles in a string array.

Function added to `ordering.go`:
- `GetCyclesForTable()`: This function takes in a table name string and
  attempts to find if the table is in any cycles.
  If the table name doesn't exist in the database, nil along with
  `MissingTableError` is returned. Otherwise, `DFS` is run. If `DFS`
  does find cycles, it is processed by `getRelevantCycles()` before
  being returned alongside nil as the error.

 Other Changes:
 - Updated documentation
 - Updated the omni test case (case 9)
 - Command Code was added