Task 23: Add CLI option to output SQL script
This commit allows the SQL code responsible for fixing cycles to be
output to a file. This is done by adding a new flag `---output, -o`

This change affects the two commands `schema` and `table` in the
`validate` suit of commands. In order to accommodate this new change, new
functions were created to simplify certain processes.

A new command in `utils` called `WriteQueriesToFile()` is responsible
for writing the array of queries to a given path. This function also
handles all of the path validation, returning any errors it encounters.
This function effectively deals with anything that has a `--output, -o`
flag enabled.

To ensure that `WriteQueriesToFile()` behaves as intended, unit tests
for it were created in the new file `utils_test.go`.

Another command was added to the `Ordering` struct called
`GetSuggestionQueriesForCycles()`. This function is identical to the
`GetSuggestionQueries()` function except it only gives suggestions based
on the cycles it was given rather then all of the cycles in the schema.
In some cases, calling this function is identical to calling
`GetSuggestionQueries()`

To help with cohesion, most of the global private variables in each
command were moved over to the `validate.go` file. This was because the
two commands (`schema` and `table`) needed the same set of flags and
using synonyms was a waste of time. New private functions were created for the
`schema` and `table` commands. These functions are `addFlags()` and
`handleCmdFlags()`. This was because there were similar processes in the
two functions that could be turned into a general function.

`addFlags()` adds the following flags to the given command `run`,
`suggestions`, `verbose` and `output`. It also makes `suggestion` and
`run` mutually exclusive and `suggestion` and `output` mutually
inclusive through the `PreRunE` function. This is so that the commands
have the same behavior when it comes to their shared flags.

`handleCmdFlags()` outlines the behavior for the commands `schema` and
`table` since they have identical logic for a certain portion of their
code. This ensures that the two commands have the same behavior when it
comes to processing their flags.

Other Changes:
- Added function `ResolveGivenCycles()` to be used in tandem with the
  `GetCyclesForTable()` function. This is to give users the option to
  resolve a subset of cycles without using the CLI
- `writeToFile()` private function created to actually write input to
  a file
- updating command descriptions